requirements.txt
# Core Dependencies
python-dotenv>=0.19.0
requests>=2.26.0
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=0.24.2
tensorflow>=2.6.0
torch>=1.9.0

# Web Framework
flask>=2.0.1
flask-restful>=0.3.9
flask-sqlalchemy>=2.5.1
flask-login>=0.5.0
flask-cors>=3.0.10
gunicorn>=20.1.0

# Database
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
redis>=3.5.3

# Indian Brokers
kiteconnect>=4.1.0
smartapi-python>=1.3.0
upstox-api>=2.0.3
py5paisa>=0.2.0
iifl-api>=1.0.0

# Technical Analysis
ta-lib>=0.4.21
pandas-ta>=0.3.14b0
mplfinance>=0.12.7a17
yfinance>=0.1.63

# AI/ML
transformers>=4.9.2
xgboost>=1.4.2
lightgbm>=3.2.1
catboost>=0.26.1
prophet>=1.0.1

# Data Processing
beautifulsoup4>=4.9.3
lxml>=4.6.3
html5lib>=1.1
aiohttp>=3.7.4
websockets>=9.1

# Security
cryptography>=3.4.7
jwt>=1.3.1
bcrypt>=3.2.0
passlib>=1.7.4

# Monitoring & Logging
prometheus-client>=0.11.0
grafana-api>=1.0.3
python-logging-loki>=0.3.1
elasticsearch>=7.14.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1
pytest-mock>=3.6.1

# Documentation
sphinx>=4.1.2
mkdocs>=1.2.2
mkdocs-material>=7.2.6

# Development
black>=21.7b0
flake8>=3.9.2
mypy>=0.910
isort>=5.9.3
